graphQL/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.1
httpx==0.27.2
strawberry-graphql==0.258.0
python-dotenv==1.0.0